// frontend/src/api/model/githubCom4H1RZooraInternalDomainJoinLiveRoomResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainLiveRoom } from './githubCom4H1RZooraInternalDomainLiveRoom';

export interface GithubCom4H1RZooraInternalDomainJoinLiveRoomResponse {
  chat_id?: string;
  livekit_url?: string;
  room?: GithubCom4H1RZooraInternalDomainLiveRoom;
  token?: string;
}